Refactor ARM client impl
